+++ ServerData.cs
@@ -96,5 +96,5 @@ public record RouteInfo(string Route, string Method);
 
 public static class APIVersion
 {
-    public const string Version = "1.2.9";
+    public const string Version = "1.3.0";
 }